[REFACTOR] Feed 로직을 Feed 패키지로 이전 및 getOwnedFeedOrException 인자 잘못 작성된 부분 수정
