+++ src/main/java/org/websoso/WSSServer/auth/validator/FeedAccessValidator.java
@@ -1,14 +1,14 @@
 package org.websoso.WSSServer.auth.validator;
 
-import static org.websoso.WSSServer.exception.error.CustomFeedError.BLOCKED_USER_ACCESS;
-import static org.websoso.WSSServer.exception.error.CustomFeedError.FEED_NOT_FOUND;
-import static org.websoso.WSSServer.exception.error.CustomFeedError.HIDDEN_FEED_ACCESS;
+import static org.websoso.WSSServer.feed.exception.CustomFeedError.BLOCKED_USER_ACCESS;
+import static org.websoso.WSSServer.feed.exception.CustomFeedError.FEED_NOT_FOUND;
+import static org.websoso.WSSServer.feed.exception.CustomFeedError.HIDDEN_FEED_ACCESS;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.websoso.WSSServer.feed.domain.Feed;
 import org.websoso.WSSServer.user.domain.User;
-import org.websoso.WSSServer.exception.exception.CustomFeedException;
+import org.websoso.WSSServer.feed.exception.CustomFeedException;
 import org.websoso.WSSServer.feed.repository.FeedRepository;
 import org.websoso.WSSServer.user.service.BlockService;
