--- src/main/java/org/websoso/WSSServer/dto/feed/UserFeedGetResponse.java
@@ -1,142 +1,0 @@
-package org.websoso.WSSServer.dto.feed;
-
-import org.websoso.WSSServer.feed.domain.Feed;
-import org.websoso.WSSServer.novel.domain.Novel;
-import org.websoso.WSSServer.library.domain.UserNovel;
-import org.websoso.WSSServer.util.TimeFormatUtil;
-
-public record UserFeedGetResponse(
-        Long feedId,
-        String feedContent,
-        String createdDate,
-        Boolean isSpoiler,
-        Boolean isModified,
-        Boolean isLiked,
-        Integer likeCount,
-        Integer commentCount,
-        Long novelId,
-        String title,
-        Float novelRating,
-        Long novelRatingCount,
-        Boolean isPublic,
-        String genre,
-        Float userNovelRating,
-        String thumbnailUrl,
-        Integer imageCount,
-        Float feedWriterNovelRating
-) {
-
-    public static UserFeedGetResponse of(Feed feed, Novel novel, Long visitorId, String thumbnailUrl,
-                                         Integer imageCount) {
-        return of(
-                feed,
-                novel,
-                visitorId,
-                thumbnailUrl,
-                imageCount,
-                visitorId != null && feed.getLikes().stream()
-                        .anyMatch(like -> like.getUserId().equals(visitorId)),
-                feed.getLikes().size(),
-                feed.getComments().size()
-        );
-    }
-
-    public static UserFeedGetResponse of(Feed feed, Novel novel, Long visitorId, String thumbnailUrl,
-                                         Integer imageCount, Boolean isLiked, Integer likeCount,
-                                         Integer commentCount) {
-        boolean isModified = !feed.getCreatedDate().equals(feed.getModifiedDate());
-        Long novelRatingCount = getNovelRatingCount(novel);
-        Float novelRating = getNovelRating(novel, novelRatingCount);
-        String genreName = getNovelGenreName(novel);
-        Float userNovelRating = getUserNovelRating(novel, visitorId);
-        Float feedWriterNovelRating = getFeedWriterNovelRating(novel, feed.getUser().getUserId());
-
-        return new UserFeedGetResponse(
-                feed.getFeedId(),
-                feed.getFeedContent(),
-                TimeFormatUtil.formatRelativeDateTime(feed.getCreatedDate()),
-                feed.getIsSpoiler(),
-                isModified,
-                isLiked,
-                likeCount,
-                commentCount,
-                novel == null ?
-                        null : novel.getNovelId(),
-                novel == null ?
-                        null : novel.getTitle(),
-                novelRating,
-                novelRatingCount,
-                feed.getIsPublic(),
-                genreName,
-                userNovelRating,
-                thumbnailUrl,
-                imageCount,
-                feedWriterNovelRating
-        );
-    }
-
-    private static Long getNovelRatingCount(Novel novel) {
-        if (novel == null) {
-            return null;
-        }
-        return novel.getUserNovels()
-                .stream()
-                .filter(userNovel -> userNovel.getUserNovelRating() != 0.0f)
-                .count();
-    }
-
-    private static Float getNovelRatingSum(Novel novel) {
-        if (novel == null) {
-            return null;
-        }
-
-        return (float) novel.getUserNovels()
-                .stream()
-                .filter(userNovel -> userNovel.getUserNovelRating() != 0.0f)
-                .mapToDouble(UserNovel::getUserNovelRating)
-                .sum();
-    }
-
-    private static Float getNovelRating(Novel novel, Long novelRatingCount) {
-        if (novel == null) {
-            return null;
-        }
-        return novelRatingCount > 0
-                ? Math.round(getNovelRatingSum(novel) / novelRatingCount * 10) / 10.0f
-                : 0.0f;
-    }
-
-    private static String getNovelGenreName(Novel novel) {
-        if (novel == null) {
-            return null;
-        }
-
-        return novel.getNovelGenres().get(0).getGenre().getGenreName();
-    }
-
-    private static Float getUserNovelRating(Novel novel, Long visitorId) {
-        if (novel == null) {
-            return null;
-        }
-
-        return novel.getUserNovels()
-                .stream()
-                .filter(userNovel -> userNovel.getUser().getUserId().equals(visitorId))
-                .findFirst()
-                .map(UserNovel::getUserNovelRating)
-                .orElse(null);
-    }
-
-    private static Float getFeedWriterNovelRating(Novel novel, Long feedWriterId) {
-        if (novel == null) {
-            return null;
-        }
-
-        return novel.getUserNovels()
-                .stream()
-                .filter(userNovel -> userNovel.getUser().getUserId().equals(feedWriterId))
-                .findFirst()
-                .map(UserNovel::getUserNovelRating)
-                .orElse(null);
-    }
-}
