+++ src/main/java/org/websoso/WSSServer/feed/application/FeedFindApplication.java
@@ -9,8 +9,8 @@ import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.websoso.WSSServer.domain.common.SortCriteria;
-import org.websoso.WSSServer.dto.feed.UserFeedGetResponse;
-import org.websoso.WSSServer.dto.feed.UserFeedsGetResponse;
+import org.websoso.WSSServer.feed.controller.dto.UserFeedGetResponse;
+import org.websoso.WSSServer.feed.controller.dto.UserFeedsGetResponse;
 import org.websoso.WSSServer.dto.novel.NovelGetResponseFeedTab;
 import org.websoso.WSSServer.feed.service.FeedLikeService;
 import org.websoso.WSSServer.feed.service.FeedQueryService;
@@ -18,9 +18,9 @@ import org.websoso.WSSServer.novel.service.GenreServiceImpl;
 import org.websoso.WSSServer.user.domain.AvatarProfile;
 import org.websoso.WSSServer.domain.Genre;
 import org.websoso.WSSServer.domain.common.FeedGetOption;
-import org.websoso.WSSServer.dto.feed.FeedGetResponse;
-import org.websoso.WSSServer.dto.feed.FeedInfo;
-import org.websoso.WSSServer.dto.feed.FeedsGetResponse;
+import org.websoso.WSSServer.feed.controller.dto.FeedGetResponse;
+import org.websoso.WSSServer.feed.controller.dto.FeedInfo;
+import org.websoso.WSSServer.feed.controller.dto.FeedsGetResponse;
 import org.websoso.WSSServer.dto.popularFeed.PopularFeedsGetResponse;
 import org.websoso.WSSServer.dto.user.UserBasicInfo;
 import org.websoso.WSSServer.feed.domain.Feed;
