+++ src/main/java/org/websoso/WSSServer/feed/controller/FeedController.java
@@ -20,17 +20,17 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
 import org.websoso.WSSServer.domain.common.SortCriteria;
-import org.websoso.WSSServer.dto.feed.UserFeedsGetResponse;
+import org.websoso.WSSServer.feed.controller.dto.UserFeedsGetResponse;
 import org.websoso.WSSServer.dto.novel.NovelGetResponseFeedTab;
 import org.websoso.WSSServer.feed.application.FeedFindApplication;
 import org.websoso.WSSServer.domain.common.FeedGetOption;
-import org.websoso.WSSServer.dto.feed.FeedCreateRequest;
-import org.websoso.WSSServer.dto.feed.FeedCreateResponse;
-import org.websoso.WSSServer.dto.feed.FeedGetResponse;
-import org.websoso.WSSServer.dto.feed.FeedImageCreateRequest;
-import org.websoso.WSSServer.dto.feed.FeedImageUpdateRequest;
-import org.websoso.WSSServer.dto.feed.FeedUpdateRequest;
-import org.websoso.WSSServer.dto.feed.FeedsGetResponse;
+import org.websoso.WSSServer.feed.controller.dto.FeedCreateRequest;
+import org.websoso.WSSServer.feed.controller.dto.FeedCreateResponse;
+import org.websoso.WSSServer.feed.controller.dto.FeedGetResponse;
+import org.websoso.WSSServer.feed.controller.dto.FeedImageCreateRequest;
+import org.websoso.WSSServer.feed.controller.dto.FeedImageUpdateRequest;
+import org.websoso.WSSServer.feed.controller.dto.FeedUpdateRequest;
+import org.websoso.WSSServer.feed.controller.dto.FeedsGetResponse;
 import org.websoso.WSSServer.dto.popularFeed.PopularFeedsGetResponse;
 import org.websoso.WSSServer.feed.application.FeedLikeApplication;
 import org.websoso.WSSServer.feed.application.FeedManagementApplication;
