+++ src/main/java/org/websoso/WSSServer/feed/controller/dto/FeedCreateResponse.java
@@ -1,4 +1,4 @@
-package org.websoso.WSSServer.dto.feed;
+package org.websoso.WSSServer.feed.controller.dto;
 
 import org.websoso.WSSServer.feed.domain.FeedImage;
