+++ src/main/java/org/websoso/WSSServer/feed/controller/dto/UserFeedGetResponse.java
@@ -1,0 +1,24 @@
+package org.websoso.WSSServer.feed.controller.dto;
+
+public record UserFeedGetResponse(
+        Long feedId,
+        String feedContent,
+        String createdDate,
+        Boolean isSpoiler,
+        Boolean isModified,
+        Boolean isLiked,
+        Integer likeCount,
+        Integer commentCount,
+        Long novelId,
+        String title,
+        Float novelRating,
+        Long novelRatingCount,
+        Boolean isPublic,
+        String genre,
+        Float userNovelRating,
+        String thumbnailUrl,
+        Integer imageCount,
+        Float feedWriterNovelRating
+) {
+
+}
