+++ src/main/java/org/websoso/WSSServer/feed/exception/CustomFeedException.java
@@ -1,8 +1,7 @@
-package org.websoso.WSSServer.exception.exception;
+package org.websoso.WSSServer.feed.exception;
 
 import lombok.Getter;
 import org.websoso.common.exception.AbstractCustomException;
-import org.websoso.WSSServer.exception.error.CustomFeedError;
 
 @Getter
 public class CustomFeedException extends AbstractCustomException {
