--- src/main/java/org/websoso/WSSServer/feed/repository/FeedCategoryRepository.java
@@ -1,29 +1,0 @@
-package org.websoso.WSSServer.feed.repository;
-
-import java.util.List;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Slice;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.stereotype.Repository;
-import org.websoso.WSSServer.feed.domain.Category;
-import org.websoso.WSSServer.feed.domain.Feed;
-import org.websoso.WSSServer.feed.domain.FeedCategory;
-
-@Repository
-public interface FeedCategoryRepository extends JpaRepository<FeedCategory, Long>, FeedCategoryCustomRepository {
-
-    List<FeedCategory> findByFeed(Feed feed);
-
-    void deleteByCategoryAndFeed(Category category, Feed feed);
-
-    @Query(value = "SELECT fc.feed FROM FeedCategory fc "
-            + "WHERE fc.category = :category "
-            + "AND (:lastFeedId = 0 OR fc.feed.feedId < :lastFeedId) "
-            + "AND fc.feed.isHidden = false "
-            + "AND (:lastFeedId IS NULL "
-            + "OR fc.feed.user.userId NOT IN (SELECT b.blockedId FROM Block b WHERE b.blockingId = :userId)) "
-            + "ORDER BY fc.feed.feedId DESC")
-    Slice<Feed> findFeedsByCategory(Category category, Long lastFeedId, Long userId, PageRequest pageRequest);
-
-}
